+++ src/com/geoscope/GeoEye/Space/TypesSystem/CoComponent/ObjectModel/GeoMonitoredObject1/DEVICE/LANModule/TLANConnectionClient.java
@@ -170,7 +170,7 @@ public class TLANConnectionClient extends TCancelableThread {
 		System.arraycopy(BA,0, LoginBuffer,0, BA.length);
 		BA = TDataConverter.ConvertInt32ToLEByteArray((int)Repeater.UserID);
 		System.arraycopy(BA,0, LoginBuffer,2, BA.length);
-		BA = TDataConverter.ConvertInt32ToLEByteArray(Repeater.idGeographServerObject);
+		BA = TDataConverter.ConvertInt32ToLEByteArray((int)Repeater.idGeographServerObject);
 		System.arraycopy(BA,0, LoginBuffer,10, BA.length);
 		short CRC = Buffer_GetCRC(LoginBuffer, 10,8);
 		BA = TDataConverter.ConvertInt16ToLEByteArray(CRC);
