+++ src/com/geoscope/GeoEye/Space/TypesSystem/Visualizations/TileImagery/TTileImageryServer.java
@@ -6,7 +6,7 @@ import com.geoscope.GeoEye.Space.Server.TGeoScopeSpaceDataServer;
 
 public class TTileImageryServer extends TGeoScopeSpaceDataServer {
 
-	public TTileImageryServer(Context pcontext, String pServerAddress, int pServerPort, int pUserID, String pUserPassword) {
+	public TTileImageryServer(Context pcontext, String pServerAddress, int pServerPort, long pUserID, String pUserPassword) {
 		super(pcontext, pServerAddress,pServerPort, pUserID,pUserPassword);
 	}
 }
